Disallow out of order buffer
I would propose changing this value to 0 or at least making it configurable. 

On wireless networks, the packets may not arrive in order. On a poor connection, this misordering can be greatly mitigated by actually sorting the packets. Having 100 packets out of order, on an average 15fps, means that you don't have any ordering on a 6 second or lower buffer. 

It also means that any ordering is only accurate to 6 seconds which is unacceptable for "real time" streaming.